% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_string.R
\name{clean_string}
\alias{clean_string}
\title{clean a string}
\usage{
clean_string(string)
}
\arguments{
\item{string}{a string to be cleared}
}
\value{
a string without white space, non alpha-numeric characters are replaced by a hash tag
}
\description{
clean a string
}
\examples{
clean_string("  Giraffen_2b")
}
