% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_strings_in_df.R
\name{clean_strings_in_df}
\alias{clean_strings_in_df}
\title{clean all characters in a tibble/ data frame}
\usage{
clean_strings_in_df(data)
}
\arguments{
\item{data}{a tibble/ dataframe}
}
\value{
a tibble / dataframe where whitespace is removed and empty characters are NAs
}
\description{
clean all characters in a tibble/ data frame
}
\examples{
print(sailor_students)
sailor_students_cleaned <- clean_strings_in_df(sailor_students)
print(sailor_students_cleaned)
}
