% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_birthdaymonth.R
\name{inspect_birthday}
\alias{inspect_birthday}
\title{inspect borhtday and month components of a SGIC}
\usage{
inspect_birthday(code, start, end)
}
\arguments{
\item{code}{SGIC}

\item{start}{start index of birthday and birthmonth}

\item{end}{end index of birthday and birthmonth}
}
\value{
Boolean
}
\description{
inspect borhtday and month components of a SGIC
}
