% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_birthmonth.R
\name{inspect_birthmonth}
\alias{inspect_birthmonth}
\title{inspect birthmonth component of a string}
\usage{
inspect_birthmonth(code, start, end)
}
\arguments{
\item{code}{a string containing the SGIC}

\item{start}{start position of birthmonth}

\item{end}{end position of birthmonth}
}
\value{
a logical value `TRUE` if the substring represents a valid birthmonth, otherwise `FALSE`.
}
\description{
This function checks if a substring of a given code represents a valid birthmonth in the code
}
\examples{
inspect_birthmonth("DEF2802", 6, 7) # TRUE
inspect_birthmonth("GHI3033", 6, 7) # FALSE
}
