% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpsect_codestructure.R
\name{inspect_codestructure}
\alias{inspect_codestructure}
\title{inspect code structure}
\usage{
inspect_codestructure(code, pattern)
}
\arguments{
\item{code}{a string containing a SGIC}

\item{pattern}{an expected pattern / regex}
}
\value{
a logical value `TRUE` if the code matches the expected pattern, otherwise `FALSE`
}
\description{
inspect code structure
}
\examples{
inspect_codestructure("ABC1234", "^[A-Za-z]{3}[0-9]{4}$") #TRUE
}
