% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_letters.R
\name{inspect_letters}
\alias{inspect_letters}
\title{inspect number of letters in a string}
\usage{
inspect_letters(code, expected_count)
}
\arguments{
\item{code}{string containing a SGIC}

\item{expected_count}{expected number of letters}
}
\value{
a logical value `TRUE` if the string contains the expected number of letters, otherwise `FALSE`.
}
\description{
inspect number of letters in a string
}
\examples{
inspect_letters("DEF2802", 3) # TRUE
inspect_letters("GH6702", 3) # FALSE
}
