% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_numberid.R
\name{inspect_numberid}
\alias{inspect_numberid}
\title{Inspect if a number has the expected length}
\usage{
inspect_numberid(number, expected_length)
}
\arguments{
\item{number}{A numeric value.}

\item{expected_length}{An integer specifying the expected number of digits.}
}
\value{
A logical value: `TRUE` if `number` has the expected length and consists only of digits, otherwise `FALSE`.
}
\description{
Check whether a given numeric value has the expected number of digits.
}
\examples{
inspect_numberid(12345, 5)  # TRUE - 5 digits
inspect_numberid(1234, 5)    # FALSE - 4 digits
}
