% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_valinvec.R
\name{inspect_valinvec}
\alias{inspect_valinvec}
\title{Inspect if a value is in a specified recode map}
\usage{
inspect_valinvec(value, recode_map)
}
\arguments{
\item{value}{A single value to inspect}

\item{recode_map}{A named vector where the names represent the keys to check against}
}
\value{
a logical value `TRUE` if the value is a key in the recode_map, otherwise `FALSE`
}
\description{
Inspect if a value is in a specified recode map
}
\examples{
recode_map <- c(male = "M", female = "F")
inspect_valinvec("male", recode_map) #TRUE
inspect_valinvec("other", recode_map) #FALSE
}
