% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_keydata.R
\name{load_keydata}
\alias{load_keydata}
\title{load and filter data from an rds-file}
\usage{
load_keydata(filepath, selection = NULL, ...)
}
\arguments{
\item{filepath}{path to the rds-file}

\item{selection}{a list of variables to dplyr::select()}

\item{...}{filter conditions to apply on the data}
}
\value{
a filtered tibble
}
\description{
load and filter data from an rds-file
}
\examples{
\dontrun{test <- load_keydata(filepath = "my/path/keydata.rds",
selection = c(year, GUID, schoolnumber, grade, class_designation, sex, my_code),
Quelle == 2324,
Klassenstufe == 5)}

}
