% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge_string.R
\name{purge_string}
\alias{purge_string}
\title{Replace non-alphanumeric characters and remove spaces and tabs in character columns of a tibble or data frame}
\usage{
purge_string(data, ..., replacement = "#", keep = "")
}
\arguments{
\item{data}{A dataframe or tibble containing columns to be processed}

\item{...}{variables to prepare for matching}

\item{replacement}{The character or string to replace non-alphanumeric characters with. Default is "#"}

\item{keep}{characters that shall not be replaced}
}
\value{
A modified dataframe/ tibble with cleaned character columns
}
\description{
Replace non-alphanumeric characters and remove spaces and tabs in character columns of a tibble or data frame
}
\examples{
print(sailor_students)
sailor_students_cleaned <- 
purge_string(sailor_students, sgic, school, class, gender, replacement = "#")
print(sailor_students_cleaned)
}
