% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_valinvec.R
\name{recode_valinvec}
\alias{recode_valinvec}
\title{recode a variable in a tibble / dataframe}
\usage{
recode_valinvec(data, var, recode_map, new_var)
}
\arguments{
\item{data}{a tibble / dataframe}

\item{var}{a variable to be recoded}

\item{recode_map}{a names vector specifying the recode map}

\item{new_var}{name of the new variable to be generated with the recoded values from 'recode_map'}
}
\value{
a tibble  / dataframe with the new variable added
}
\description{
recode a variable in a tibble / dataframe
}
\examples{
print(sailor_students)
recode_map_gender <- c("Female" = "F", "Male" = "M", "Other" = "X")
sailor_students_recoded <- 
recode_valinvec(sailor_students, gender, recode_map_gender, recode_gender)
print(sailor_students_recoded)
}
