% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_nonalnum_in_df.R
\name{replace_nonalnum_in_df}
\alias{replace_nonalnum_in_df}
\title{Replace non-alphanumeric characters and remove spaces and tabs in character columns of a tibble or data frame, umlauts and eszett remain unchanged}
\usage{
replace_nonalnum_in_df(data, replacement = "#")
}
\arguments{
\item{data}{A dataframe or tibble containing columns to be processed}

\item{replacement}{The character or string to replace non-alphanumeric characters with. Default is "#"}
}
\value{
A modified dataframe/ tibble with cleaned character columns
}
\description{
Replace non-alphanumeric characters and remove spaces and tabs in character columns of a tibble or data frame, umlauts and eszett remain unchanged
}
\examples{
data <- tibble::tibble(
  name = c("Alice", "Bob", "Charlie"),
  description = c("Hello! 123 and special characters #$\%.   Tabs:     ",
                  "This one has umlauts and spec!al char$cters...",
                  "Another sentence with @#$\% special! characters and no numbers.")
)
data_modified <- replace_nonalnum_in_df(data, "#")
}
