% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_dupes.R
\name{report_dupes}
\alias{report_dupes}
\title{duplicate cases based on specified variables}
\usage{
report_dupes(data, ...)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{...}{variable names to check for duplicates}
}
\value{
a tibble containing only duplicate cases
}
\description{
duplicate cases based on specified variables
}
\examples{
data <- tibble::tibble(
  id = 1:6,
 name = c("Alice", "Bob", "Alice", "David", "Alice", "Bob"),
 age = c(25, 30, 25, 40, 25, 30)
)

duplicates <- report_dupes(data, name, age)
}
