% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_invalid_birthday.R
\name{report_invalid_birthday}
\alias{report_invalid_birthday}
\title{report invalid birthday components of strings}
\usage{
report_invalid_birthday(data, code, start, end)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{code}{a string variable containing a SGIC}

\item{start}{start position of birthday}

\item{end}{en position of birthday}
}
\value{
a tibble containing only SGICs with invalid birthdays
}
\description{
report invalid birthday components of strings
}
\examples{

# student data
data <- tibble(
code = c("APE1601", "TBG3568", "DFR2508")
)

# report invalid birthdays in code
invalid_birthdays <- report_invalid_birthdays(data, "code", 4, 5)

# show results
print(invalid_birthdays)

}
