% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_invalid_birthdaymonth.R
\name{report_invalid_birthdaymonth}
\alias{report_invalid_birthdaymonth}
\title{report strings with invalid birthdate entries}
\usage{
report_invalid_birthdaymonth(data, code)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{code}{a string variable containing a SGIC}
}
\value{
a tibble containing only SGICs with invalid birthdates
}
\description{
report strings with invalid birthdate entries
}
\examples{

# student data
data <- tibble::tibble(
code = c("APE1601", "TBG3568", "DFR2508")
)

# report invalid birthdaymonths in code
invalid_birthdaymonths <- report_invalid_birthdaymonth(data, "code")

# show results
print(invalid_birthdaymonths)

}
