% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_invalid_birthmonth.R
\name{report_invalid_birthmonth}
\alias{report_invalid_birthmonth}
\title{report strings with invalid birthmonth entries}
\usage{
report_invalid_birthmonth(data, code)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{code}{a string variable containing a SGIC}
}
\value{
a tibble containing only SGICs with invalid birthmonth
}
\description{
report strings with invalid birthmonth entries
}
\examples{

# student data
data <- tibble::tibble(
code = c("APE1601", "TBG3568", "DFR2508")
)

# report invalid birthmonths in code
invalid_birthmonths <- report_invalid_birthmonth(data, "code")

# show results
print(invalid_birthmonths)

}
