% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_invalid_codestructure.R
\name{report_invalid_codestructure}
\alias{report_invalid_codestructure}
\title{report invalid code structure}
\usage{
report_invalid_codestructure(data, code, pattern)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{code}{a string containing a SGIC}

\item{pattern}{an expected pattern / regex}
}
\value{
a tibble containing only SGICs with incalid code structure
}
\description{
report invalid code structure
}
\examples{
# student data
data <- tibble(
code = c("AE1601", "TBG368", "DFR2508")
)

# report invalid code structure
invalid_codestructures <- report_invalid_codestructure(data, "code", "^[A-Za-z]{3}[0-9]{4}$")

# show results
print(invalid_codestructures)
}
