% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_invalid_letters.R
\name{report_invalid_letters}
\alias{report_invalid_letters}
\title{report invalid number of letters in a string}
\usage{
report_invalid_letters(data, code, expected_count)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{code}{a string variable containing a SGIC}

\item{expected_count}{expected number of letters}
}
\value{
a tibble containing only SGICs with invalid number of letters
}
\description{
report invalid number of letters in a string
}
\examples{

# student data
data <- tibble(
code = c("APE1601", "TBGR568", "DFR2508")
)

# report invalid number of letters in code
invalid_letters <- report_invalid_letters(data, "code")

# show results
print(invalid_letters)


}
