% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_invalid_schoolnumber.R
\name{report_invalid_schoolnumber}
\alias{report_invalid_schoolnumber}
\title{report invalid schoolnumbers}
\usage{
report_invalid_schoolnumber(data, schoolnumber, expected_length)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{schoolnumber}{a numeric value}

\item{expected_length}{expected length of the numeric value}
}
\value{
a tibble containing only schoolnumbers not equal to the expected length
}
\description{
report invalid schoolnumbers
}
\examples{

# schoolnumbers
data <- tibble::tibble(
schoolnumber = c(12345, 54321, 123)
)

# report invalid schoolnumbers
invalid_schoolnumbers <- report_invalid_schoolnumber(data, "schoolnumber", 5)

# show results
print(invalid_schoolnumbers)

}
