% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sailor_keys}
\alias{sailor_keys}
\title{key data on students from the sailor moon universe}
\format{
`sailor_keys`
A tibble with 12 rows and 6 columns:
\describe{
  \item{schoolyear}{schoolyear}
  \item{guid}{hexadecimal ID number}
  \item{name, birthday, sex}{student information}
  \item{school, schoolnumber, class, grade_level}{school information}
  \item{sgic1, sgic2, sgic3}{subject generated ID}
}
}
\usage{
sailor_keys
}
\description{
A fictional key data set.
}
\keyword{datasets}
