% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sailor_students}
\alias{sailor_students}
\title{assessment data on students from the sailor moon universe}
\format{
## `sailor_students`
A data frame with 12 rows and 6 columns:
\describe{
  \item{sgic}{Subjec generated ID}
  \item{school}{schoolnumber}
  \item{class}{class designation}
  \item{gender}{gender}
  \item{testscore_lannguage, testscore_calculus}{testscores}
}
}
\usage{
sailor_students
}
\description{
A fictional assessment dataset.
}
\keyword{datasets}
