% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summon_data.R
\name{summon_data}
\alias{summon_data}
\title{load and filter data from an rds-file}
\usage{
summon_data(filepath, selection = NULL, ...)
}
\arguments{
\item{filepath}{path to the rds-file}

\item{selection}{a list of variables to dplyr::select()}

\item{...}{filter conditions to apply on the data}
}
\value{
a filtered tibble
}
\description{
load and filter data from an rds-file
}
\examples{

\dontrun{test <- summon_data(filepath = "my/path/keydata.rds",
selection = c(year, GUID, schoolnumber, grade, class_designation, sex, sgic),
year == 2324,
class_designation == 5)}

}
