% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_strings_in_df.R
\name{clean_strings_in_df}
\alias{clean_strings_in_df}
\title{clean all characters in a tibble/ data frame}
\usage{
clean_strings_in_df(data)
}
\arguments{
\item{data}{a tibble/ dataframe}
}
\value{
a tibble / dataframe where whitespace is removed and empty characters are NAs
}
\description{
clean all characters in a tibble/ data frame
}
\examples{

#example student data
data <- tibble(
name = c(" Oli Bock", "Oli Gorth   ", "  Oli Hackbarth "),
testscore = c(125, 195, 345),
sgic = c("OEB_1345", "LKR 29 854", " R T SSS 1356"))

#apply function
clean_strings_in_df(data)

print results
print(data)

}
