% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_birthday.R
\name{inspect_birthday}
\alias{inspect_birthday}
\title{inspect if a string contains only one number-component which represents a valid birthday}
\usage{
inspect_birthday(code)
}
\arguments{
\item{code}{a string containing the SGIC}
}
\value{
a logical value `TRUE` if the string contains only one valid birthday-component, otherwise `FALSE`.
}
\description{
This function checks if a string contains only one number-component which represents a valid birthday
}
\examples{
inspect_birthday("DEF66") # FALSE
inspect_birthday("GHI02") # TRUE
}
