% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_birthday.R
\name{inspect_birthday}
\alias{inspect_birthday}
\title{inspect birthday component of a string}
\usage{
inspect_birthday(code)
}
\arguments{
\item{code}{a string containing the SGIC}
}
\value{
a logical value `TRUE` if the substring represents a valid birthday, otherwise `FALSE`.
}
\description{
This function checks if a substring of a given code represents a valid birthday in the code
}
\examples{
inspect_birthday("DEF2802") # TRUE
inspect_birthday("GHI6702") # FALSE
}
