% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_birthdaymonth.R
\name{inspect_birthdaymonth}
\alias{inspect_birthdaymonth}
\title{inspect birthday and -month component of a string}
\usage{
inspect_birthdaymonth(code)
}
\arguments{
\item{code}{a string containing the SGIC}
}
\value{
a logical value `TRUE` if the string hat only one number-component representing a valid birthday and -month, otherwise `FALSE`.
}
\description{
This function checks if a string contains only one number-component representing a valid combination of birthday followed by month
}
\examples{
inspect_birthdaymonth("DEF2802") # TRUE
inspect_birthdaymonth("GHI3002") # FALSE

}
