% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_birthdaymonth.R
\name{inspect_birthdaymonth}
\alias{inspect_birthdaymonth}
\title{inspect birthday and -month component of a string}
\usage{
inspect_birthdaymonth(code, start, end)
}
\arguments{
\item{code}{a string containing the SGIC}

\item{start}{start position of birthdate-components}

\item{end}{end position of birthdate-components}
}
\value{
a logical value `TRUE` if the substring represents a valid date, otherwise `FALSE`.
}
\description{
This function checks if a substring of a given code represents a valid birthdate in the code
}
\examples{
inspect_birthdaymonth("DEF2802", 4, 7) # TRUE
inspect_birthdaymonth("GHI3002", 4, 7) # FALSE

}
