% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_birthmonth.R
\name{inspect_birthmonth}
\alias{inspect_birthmonth}
\title{inspect if a string contains only one number-component which represents a valid birthmonth}
\usage{
inspect_birthmonth(code)
}
\arguments{
\item{code}{a string containing the SGIC}
}
\value{
a logical value `TRUE` if the string contains only one valid birthmonth-component, otherwise `FALSE`.
}
\description{
This function checks if a string contains only one number-component which represents a valid birthmonth
}
\examples{
inspect_birthmonth("DEF02") # TRUE
inspect_birthmonth("GHI33") # FALSE
}
