% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_codestructure.R
\name{inspect_codestructure}
\alias{inspect_codestructure}
\title{inspect whether a string matches an expected pattern}
\usage{
inspect_codestructure(code, pattern)
}
\arguments{
\item{code}{a string containing a SGIC}

\item{pattern}{an expected pattern / regex}
}
\value{
a logical value `TRUE` if the code matches the expected pattern, otherwise `FALSE`
}
\description{
This function checks if a string matches an expected pattern
}
\examples{
inspect_codestructure("ABC1234", "^[A-Za-z]{3}[0-9]{4}$") #TRUE
inspect_codestructure("12DBG45FG", "^[A-Za-z]{3}[0-9]{4}$") #FALSE
}
