% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_dupes.R
\name{inspect_dupes}
\alias{inspect_dupes}
\title{Inspect data for duplicate cases based on specified variables}
\usage{
inspect_dupes(data, ...)
}
\arguments{
\item{data}{A data frame / tibble containing the data to be inspected}

\item{...}{variable names to check for duplicates}
}
\value{
The original data frame with an additional logical column 'has_dupes'
}
\description{
Inspect data for duplicate cases based on specified variables
}
\examples{
print(sailor_students)
sailor_students_dupes <- inspect_dupes(sailor_students, sgic, school, class)
print(sailor_students_dupes)
}
