% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_numberid.R
\name{inspect_numericid}
\alias{inspect_numericid}
\title{inspect if a number has the expected length}
\usage{
inspect_numericid(number, expected_length)
}
\arguments{
\item{number}{a numeric value}

\item{expected_length}{expected digits of schoolnumber}
}
\value{
a logical value `TRUE` if the substring represents a valid schoolnumber, otherwise `FALSE`
}
\description{
inspect if a number has the expected length
}
\examples{
inspect_numericid(12345, 5)  # TRUE
inspect_numericid(123456, 5)  # FALSE
}
