% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpsect_schoolnumber.R
\name{inspect_schoolnumber}
\alias{inspect_schoolnumber}
\title{inspect if a schoolnumber has the expected length}
\usage{
inspect_schoolnumber(schoolnumber, expected_length)
}
\arguments{
\item{schoolnumber}{a numeric value}

\item{expected_length}{expected digits of schoolnumber}
}
\value{
a logical value `TRUE` if the substring represents a valid birthmonth, otherwise `FALSE`
}
\description{
inspect if a schoolnumber has the expected length
}
\examples{
inspect_schoolnumber(12345, 5)  # TRUE
inspect_schoolnumber(123456, 5)  # FALSE
}
