% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_nonalnum_in_df.R
\name{replace_nonalnum_in_df}
\alias{replace_nonalnum_in_df}
\title{Replace non-alphanumeric characters and remove spaces and tabs in character columns of a tibble or data frame, umlauts and eszett remain unchanged, empty strings get replaced}
\usage{
replace_nonalnum_in_df(data, ..., replacement = "#")
}
\arguments{
\item{data}{A dataframe or tibble containing columns to be processed}

\item{...}{variables to prepare for matching}

\item{replacement}{The character or string to replace non-alphanumeric characters with. Default is "#"}
}
\value{
A modified dataframe/ tibble with cleaned character columns
}
\description{
Replace non-alphanumeric characters and remove spaces and tabs in character columns of a tibble or data frame, umlauts and eszett remain unchanged, empty strings get replaced
}
\examples{
print(sailor_students)
sailor_students_cleaned <- replace_nonalnum_in_df(sailor_students, sgic, school, class, gender, replacement = "#")
print(sailor_students_cleaned)
}
