% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_invalid_birthdaymonth.R
\name{report_invalid_birthdaymonth}
\alias{report_invalid_birthdaymonth}
\title{report invalid birthday and -month components of strings}
\usage{
report_invalid_birthdaymonth(data, code, start, end)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{code}{a string variable containing a SGIC}

\item{start}{start position of birthdate}

\item{end}{en position of birthdate}
}
\value{
a tibble containing only SGICs with invalid birthdates
}
\description{
report invalid birthday and -month components of strings
}
\examples{

# student data
data <- tibble(
code = c("APE1601", "TBG3568", "DFR2508")
)

# report invalid birthdaymonths in code
invalid_birthdaymonths <- report_invalid_birthdaymonth(data, "code", 6, 7)

# show results
print(invalid_birthdaymonths)

}
