% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_invalid_birthmonth.R
\name{report_invalid_birthmonth}
\alias{report_invalid_birthmonth}
\title{report invalid birthmonth components of strings}
\usage{
report_invalid_birthmonth(data, code, start, end)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{code}{a string variable containing a SGIC}

\item{start}{start position of birthmonth}

\item{end}{en position of birthmonth}
}
\value{
a tibble containing only SGICs with invalid birthmonth
}
\description{
report invalid birthmonth components of strings
}
\examples{

# student data
data <- tibble(
code = c("APE1601", "TBG3568", "DFR2508")
)

# report invalid birthmonths in code
invalid_birthmonths <- report_invalid_birthmonth(data, "code", 6, 7)

# show results
print(invalid_birthmonths)

}
