% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sailor_students}
\alias{sailor_students}
\title{assessment data on students from the sailor moon universe}
\format{
`sailor_students`
A tibble with 12 rows and 6 columns:
\describe{
  \item{sgic}{Subject generated ID}
  \item{school}{schoolnumber}
  \item{class}{class designation}
  \item{gender}{gender}
  \item{testscore_language, testscore_calculus}{testscores}
}
}
\usage{
sailor_students
}
\description{
A fictional assessment data set.
}
\keyword{datasets}
