% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summon_data.R
\name{summon_data}
\alias{summon_data}
\title{Load an RDS-file and select/filter Variables}
\usage{
summon_data(filepath, selection = NULL, ...)
}
\arguments{
\item{filepath}{path to the rds-file to be loaded.}

\item{selection}{An optional character vector of variable names to be selected from the data using `dplyr::select()`. 
If not provided, all columns from the data will be retained.}

\item{...}{Additional filtering conditions that will be applied to the data using `dplyr::filter()`.}
}
\value{
A filtered tibble containing the selected variables and rows that meet the specified filter conditions.
}
\description{
Read data from a RDS file, filter the resulting tibble based on the provided 
filter conditions and optionally specify a list of variables to select from the data. 
If the data is not already a tibble, it will be converted before any operations are performed.
}
\examples{

\dontrun{
test <- summon_data(filepath = "my/path/keydata.rds",
selection = c(year, GUID, schoolnumber, grade, class_designation, sex, sgic),
year == 2324,
class_designation == 5
)}

}
